+++ module/migration.js
@@ -1,4 +1,5 @@
 import { ArM5ePCActor } from "./actor/actor.js";
+import { ActorDataModels, ItemDataModels } from "./arm5e.js";
 import { error, log } from "./tools.js";
 
 const DEPRECATED_ITEMS = ["speciality", "distinctive", "sanctumRoom", "personality"];
@@ -315,11 +316,8 @@ export const migrateActorData = async function (actorDoc, actorItems) {
   } else {
     actor = actorDoc;
   }
-  if (CONFIG.Actor.systemDataModels[actor.type]) {
-    updateData = CONFIG.Actor.systemDataModels[actor.type].migrate(
-      actor,
-      actorDoc.items ? actorDoc.items : []
-    );
+  if (ActorDataModels[actor.type]) {
+    updateData = ActorDataModels[actor.type].migrate(actor, actorDoc.items ? actorDoc.items : []);
   }
 
   if (!actor?.flags?.arm5e) {
@@ -919,8 +917,8 @@ export const migrateItemData = async function (item) {
   } else {
     itemData = item;
   }
-  if (CONFIG.Item.systemDataModels[item.type]) {
-    return CONFIG.Item.systemDataModels[item.type].migrate(itemData);
+  if (ItemDataModels[item.type]) {
+    return ItemDataModels[item.type].migrate(itemData);
   }
   const updateData = {};
   if (_isMagicalItem(itemData)) {